# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Simpay API",
    "description": "API for Simpay that allows users to buy tickets for events",
    "version": "1.0.0"
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Bearer token"
    }
  },
  "paths": {
    "/users": {
      "post": {
        "summary": "Create user",
        "tags": ["User"],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CreateUserParams"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created"
          },
          "400": {
            "description": "Invalid params"
          },
          "409": {
            "description": "User already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "Login user",
        "tags": ["Auth"],
        "parameters": [
          {
            "name": "body",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/LoginUserParams"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User logged in",
            "schema": {
              "$ref": "#/definitions/LoginResponse"
            }
          },
          "400": {
            "description": "Invalid params"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/me/tickets": {
      "get": {
        "summary": "List authenticated user's tickets",
        "description": "Returns all tickets purchased by the authenticated user, including event details.",
        "tags": ["User"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Tickets found",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UserTicketResponse"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "CreateUserParams": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "John Doe"
        },
        "email": {
          "type": "string",
          "example": "Jhon@example.com"
        },
        "password": {
          "type": "string",
          "example": "password123"
        },
        "role": {
          "type": "string",
          "example": "customer"
        }
      }
    },
...
    "UserWithoutPassword": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "e299b4d5-c6e9-4a0a-a133-3a86f9daa3d5"
        },

        "name": {
          "type": "string",
          "example": "John Doe"
        },
        "email": {
          "type": "string",
          "example": "Jhon@example.com"
        },
        "role": {
          "type": "string",
          "example": "customer"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "example": "2027-01-01T00:00:00.000Z"
        }
      }
    },
    "LoginUserParams": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "LoginResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/UserWithoutPassword"
        }
      }
    },
...
    "EventSummary": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "e299b4d5-c6e9-4a0a-a133-3a86f9daa3d5"
        },
        "name": {
          "type": "string",
          "example": "My Event"
        },
        "date": {
          "type": "string",
          "format": "date-time",
          "example": "2027-01-01T00:00:00.000Z"
        }
      }
    },
    "UserTicketResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "example": "e299b4d5-c6e9-4a0a-a133-3a86f9daa3d5"
        },
        "quantity": {
          "type": "integer",
          "example": 2
        },
        "totalPriceInCents": {
          "type": "integer",
          "example": 200
        },
        "status": {
          "type": "string",
          "example": "paid"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time",
          "example": "2027-01-01T00:00:00.000Z"
        },
        "event": {
          "$ref": "#/definitions/EventSummary"
        }
      }
    }
  }
}